+++ src/main/kotlin/com/kos/entities/sync/rules/SyncRule.kt
@@ -1,0 +1,8 @@
+package com.kos.entities.sync.rules
+
+import com.kos.entities.domain.Entity
+import com.kos.views.Game
+
+sealed interface SyncRule {
+    suspend fun pick(game: Game): List<Entity>
+}