+++ src/main/kotlin/com/kos/sources/wow/WowEntitySynchronizer.kt
@@ -10,10 +10,10 @@ import com.kos.common.DynamicCache
 import com.kos.common.WithLogger
 import com.kos.common.error.ServiceError
 import com.kos.datacache.DataCache
-import com.kos.datacache.EntitySynchronizer
 import com.kos.datacache.repository.DataCacheRepository
 import com.kos.entities.domain.Entity
 import com.kos.entities.domain.WowEntity
+import com.kos.entities.sync.EntitySynchronizer
 import com.kos.sources.wow.staticdata.wowseason.repository.WowSeasonRepository
 import com.kos.views.Game
 import kotlinx.coroutines.channels.Channel