+++ src/test/kotlin/com/kos/datacache/DataCacheServiceTest.kt
@@ -77,11 +77,6 @@ class DataCacheServiceTest {
         }
     }
 
-    @Test
-    fun `it must not return no data and sync operation when getOrSync entity was not in the system`() {
-
-    }
-
     private suspend fun createService(dataCacheRepository: DataCacheInMemoryRepository): DataCacheService {
         val viewsRepository = ViewsInMemoryRepository()
             .withState(